move interfaces to separate modules
