--- authn/webapi/interfaces/pgsqlx/constants/constants.go
@@ -1,4 +1,0 @@
-package constants
-
-// ConnectionString -
-const ConnectionString = "postgresql://%s:%s@%s:%d/%s?sslmode=disable"
